Album model to ts
