+++ src/core/db.ts
@@ -38,7 +38,7 @@ function init(db) {
 
     Data.setStore('Page', db.collection('pages'))
     Data.setStore('Article', db.collection('articles'))
-    Data.setStore(Album, db.collection('albums'))
+    Data.setStore('Album', db.collection('albums'))
     Data.setStore('Image', db.collection('images'))
     Data.setStore('Teacher', db.collection('collective'))
     Data.setStore(Document, db.collection('documents'))
